+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityEditModel.java
@@ -225,16 +225,16 @@ public class SwingEntityEditModel extends AbstractEntityEditModel {
 		clearForeignKeyReferences(foreignKey, entities);
 		if (comboBoxModels.containsKey(foreignKey)) {
 			EntityComboBoxModel comboBoxModel = foreignKeyComboBoxModel(foreignKey);
-			Entity selectedEntity = comboBoxModel.selection().value();
+			Entity selectedEntity = comboBoxModel.selection().item().get();
 			entities.forEach(comboBoxModel.items()::removeItem);
 			if (comboBoxModel.items().visible().contains(selectedEntity)) {
-				comboBoxModel.setSelectedItem(selectedEntity);
-			}//if the null value is selected we're fine, otherwise select topmost item
-			else if (!comboBoxModel.selection().nullSelected() && comboBoxModel.getSize() > 0) {
-				comboBoxModel.setSelectedItem(comboBoxModel.getElementAt(0));
+				comboBoxModel.selection().item().set(selectedEntity);
+			}//if null is selected we're fine, otherwise select topmost item
+			else if (comboBoxModel.selection().item().get() != null && comboBoxModel.getSize() > 0) {
+				comboBoxModel.selection().item().set(comboBoxModel.getElementAt(0));
 			}
 			else {
-				comboBoxModel.setSelectedItem(null);
+				comboBoxModel.selection().item().clear();
 			}
 		}
 	}