+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/component/DefaultEntityComboBoxModel.java
@@ -373,7 +373,7 @@ final class DefaultEntityComboBoxModel implements EntityComboBoxModel {
 			if (!filterKeys.isEmpty()) {
 				filterModel.select(filterKeys.iterator().next());
 			}
-			set(selection().value());
+			set(selection().item().get());
 			filterModel.selection().item().addConsumer(this::set);
 			selection().item().addConsumer(selected -> select(filterModel, selected));
 			refresher().success().addListener(filterModel::refresh);